Refactor PermissionCheck to take Context instead of Activity
An activity is only needed when _requesting_ a permission.
This we should only do when the UI is visible, e.g., from a fragment.

Checking whether a permission is granted only requires a Context.
This is weaker (but also easier), e.g., the Context of a background service suffices.

Thus this commit separates checking and requesting permissions, so that we only
use an activity where it is really needed, and use a Context everywhere else.